spring boot 설정만 적용하는 것으로 완료함.
