--- src/main/resources/application.properties
@@ -1,20 +1,0 @@
-spring.profiles.active= local
-spring.profiles= local
-spring.basic.enabled= false
-
-spring.jpa.hibernate.dialect= net.slipp.support.jpa.Mysql5BitBooleanDialect
-spring.jpa.hibernate.ddl-auto= validate
-spring.jpa.hibernate.naming_strategy= org.hibernate.cfg.ImprovedNamingStrategy
-spring.jpa.show-sql= true
-spring.jpa.database= mysql
-
-spring.jpa.properties.hibernate.cache.use_second_level_cache= true
-spring.jpa.properties.hibernate.cache.use_query_cache= true
-spring.jpa.properties.hibernate.cache.provider_class= org.hibernate.cache.SingletonEhCacheProvider
-spring.jpa.properties.hibernate.cache.region.factory_class= org.hibernate.cache.ehcache.SingletonEhCacheRegionFactory
-
-spring.datasource.driverClassName= com.mysql.jdbc.Driver
-spring.datasource.url= jdbc:mysql://127.0.0.1/slipp?useUnicode=yes&amp;characterEncoding=UTF-8
-spring.datasource.username= slipp
-spring.datasource.password= password
-
