--- src/main/scala/net/slipp/SlippWebApplication.scala
@@ -1,7 +1,0 @@
-package net.slipp
-
-import org.springframework.boot.SpringApplication
-
-object SlippWebApplication extends App {
-  SpringApplication.run(classOf[SlippConfig])
-}
